Lazy load (vue-router)  Meatball's welcome component
