+++ public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=ad612a4277f6d583514f",
+    "/js/app.js": "/js/app.js?id=13e6c050d8538ab97649",
     "/css/app.css": "/css/app.css?id=469cf92298e3edd62899"
 }
