+++ resources/js/bootstrap.js
@@ -1,9 +1,3 @@
-// window._ = require('lodash');
-
-// try {
-//     window.Popper = require('popper.js').default;
-// } catch (e) {}
-
 window.axios = require('axios');
 
 window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
